+++ RatRaceOnline/ViewModel/CarouselViewModel.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace RatRace3
+namespace RatRace3.ViewModel
 {
     public class CarouselViewModel
     {
